updated int to longs

+++ prefix_pthreads.c
@@ -40,7 +40,7 @@ static int*arrayC;
 static int*lastPrefix;
 static int size;
 
-
+//1000000000
 int main(int argc, char* argv[])
 {
                               printf("Prefix Sum v1.0 [pthreads]\n");
@@ -192,16 +192,16 @@ int main(int argc, char* argv[])
 
 static void* pthreadCalc(void* rank){
 	long long my_rank = (long long)rank;
-	int my_start = my_rank * size / thread_count;
-	int my_end = (my_rank + 1) * size / thread_count;
+	long my_start = my_rank * size / thread_count;
+	long my_end = (my_rank + 1) * size / thread_count;
 	
-	int myS=my_end-my_start;
+	long myS=my_end-my_start;
     
     //creating a temp array for local prefix solution
 	int tempArr[myS];
-	int c=0;
+	long c=0;
     //getting partial array.
-    for(int p=my_start;p<my_end;p++){
+    for(long p=my_start;p<my_end;p++){
         //loading the corresponding part of the array into temp
         tempArr[c]=arrayA[p];
         c++;
@@ -220,14 +220,14 @@ static void* pthreadCalc(void* rank){
 
 static void* pthreadCalc2(void* rank){
 	long long my_rank = (long long)rank;
-	int my_start = my_rank * size / thread_count;
-	int my_end = (my_rank + 1) * size / thread_count;
+	long my_start = my_rank * size / thread_count;
+	long my_end = (my_rank + 1) * size / thread_count;
 	
     
     //get your index from lastPrefix and subtract last value from arrayC
-    int additionValue=lastPrefix[my_rank]-arrayC[my_end-1];
+    long additionValue=lastPrefix[my_rank]-arrayC[my_end-1];
     
-   for(int y=my_start;y<my_end;y++){
+   for(long y=my_start;y<my_end;y++){
         arrayC[y]+=additionValue;
     }
